Initial stab at converting autoseed which includes findimageoverlaps and footprintinit (eventually)
