+++ isis/src/base/apps/findimageoverlaps/findimageoverlaps.h
@@ -1,0 +1,14 @@
+#ifndef findimageoverlaps_h
+#define findimageoverlaps_h
+
+#include "UserInterface.h"
+#include "Pvl.h"
+
+using namespace std;
+using namespace Isis;
+
+namespace Isis {
+  extern void findimageoverlaps(UserInterface &ui, Pvl *log=nullptr);
+}
+
+#endif
