+++ isis/src/control/apps/autoseed/autoseed.h
@@ -1,0 +1,19 @@
+#ifndef autoseed_h
+#define autoseed_h
+
+#include "ControlNet.h"
+#include "Pvl.h"
+#include "UserInterface.h"
+
+enum SeedDomain {
+  XY,
+  SampleLine
+};
+
+namespace Isis {
+  extern void autoseed(UserInterface &ui,
+                       Pvl *log = nullptr);
+  extern void autoseed(UserInterface &ui, ControlNet *precnet = nullptr, Pvl *log=nullptr);
+}
+
+#endif
